refactor#60 drag drop state file loation
